test_requirements.txt
pytz==2018.9
pytest
pytest-dotenv
urllib3<2.0
vcrpy<3.0